UT: One approach to hiding implementation detail from interface
It might not be a good idea to expose that the tasks are stored
in a sorted list with random access, in case we want to change
that sometime in the future. This also prevents clients from
making mistakes if an item is removed or the sorting of the list
changes for some other reason.